Move reflective code to an implementation of MessageAdapter.
This will allow other implementations in the future after more refactoring of the runtime allows exposing WireOutput and WireInput.
